Sonar

+++ src/main/java/io/github/sefiraat/crystamaehistoria/utils/datatypes/PersistentUUIDDataType.java
@@ -43,7 +43,7 @@ public final class PersistentUUIDDataType implements PersistentDataType<int[], U
         Validate.notNull(ints, "The provided integer array cannot be null!");
         Validate.isTrue(ints.length == 4, "The integer array must have a length of 4.");
 
-        return new UUID(ints[0] << 32L | ints[1] & 0xFFFFFFFFL, ints[2] << 32L | ints[3] & 0xFFFFFFFFL);
+        return new UUID((long) ints[0] << 32L | ints[1] & 0xFFFFFFFFL, (long) ints[2] << 32L | ints[3] & 0xFFFFFFFFL);
     }
 
     @Nonnull
